+++ example/lib/models/list_model.dart
@@ -133,7 +133,7 @@ const List<Color> itemColors = [
 // ];
 
 //* ListModel extension
-ListModel getListModel(BuildContext context) => Pub.getModel<ListModel>();
+ListModel getListModel(BuildContext context) => Pub.model<ListModel>();
 
 Subscriber<ListModel> subListModel(CreatorFn<ListModel> create,
     {Key key, Object aspects}) {