+++ example/lib/models/my_model.dart
@@ -161,7 +161,7 @@ class MyModel extends Pub {
 }
 
 //* MyModel extension
-MyModel getMyModel(BuildContext context) => Pub.getModel<MyModel>();
+MyModel getMyModel(BuildContext context) => Pub.model<MyModel>();
 
 Subscriber<MyModel> subMyModel(CreatorFn<MyModel> create,
     {Key key, Object aspects}) {