+++ example/lib/widgets/radio_group.dart
@@ -26,7 +26,7 @@ class _RadioGroupState extends State<RadioGroup> {
   ];
 
   Future<void> _handleRadioValueChange1(String value) async {
-    final model = Pub.getModel<ListModel>();
+    final model = Pub.model<ListModel>();
     await model.changeLocale(context, value);
     setState(() {
       // model.locale.value = value; // in model.changeLocale
@@ -35,7 +35,7 @@ class _RadioGroupState extends State<RadioGroup> {
 
   @override
   Widget build(BuildContext context) {
-    final model = Pub.getModel<ListModel>();
+    final model = Pub.model<ListModel>();
     final _radioValue1 = model.locale.value;
 
     Widget panel(int index) {